refactor: moved resolvers into own foler
