+++ IndoorMapsBackend/src/resolvers/BuildingResolver.ts
@@ -3,10 +3,10 @@ import { Resolver, Query, Mutation, Arg, Ctx, InputType, Field, FieldResolver, R
 import { GraphQLError } from 'graphql'
 import { Floor as DbFloor } from '@prisma/client'
 
-import { Context } from './context.js'
-import { Building, Floor } from './Building.js'
-import { convertToGraphQLBuilding, convertToGraphQLFloor } from './utils/typeConversions.js'
-import { getUserOrThrowError } from './auth/validateUser.js'
+import { Context } from '../context.js'
+import { Building, Floor } from '../Building.js'
+import { convertToGraphQLBuilding, convertToGraphQLFloor } from '../utils/typeConversions.js'
+import { getUserOrThrowError } from '../auth/validateUser.js'
 
 @InputType()
 class BuildingUniqueInput {
